refactor(tasks): move locomotion terrain spawn [skip ci]
