+++ src/unilab/tasks/locomotion/go1/rough.py
@@ -16,9 +16,6 @@ from unilab.dr.dr_utils import build_common_reset_randomization, zero_actions
 from unilab.dtype_config import get_global_dtype
 from unilab.envs.locomotion.common import rewards
 from unilab.envs.locomotion.common.rewards import RewardContext
-from unilab.envs.locomotion.common.terrain_spawn import (
-    TerrainCurriculumCfg,
-)
 from unilab.tasks.locomotion.common.commands import (
     Commands,
     apply_heading_yaw_feedback,
@@ -33,6 +30,9 @@ from unilab.tasks.locomotion.common.height_scan import (
     init_height_scan_sensor,
     raw_height_scan_obs,
     terrain_out_of_bounds,
+)
+from unilab.tasks.locomotion.common.terrain_spawn import (
+    TerrainCurriculumCfg,
 )
 from unilab.tasks.locomotion.go1.base import ControlConfig
 from unilab.tasks.locomotion.go1.joystick import (
