+++ src/unilab/tasks/locomotion/go2w/rough.py
@@ -14,10 +14,6 @@ from unilab.dr.dr_utils import zero_actions
 from unilab.dtype_config import get_global_dtype
 from unilab.envs.locomotion.common import rewards
 from unilab.envs.locomotion.common.rewards import RewardContext
-from unilab.envs.locomotion.common.terrain_spawn import (
-    TerrainCurriculumCfg,
-    TerrainSpawnManager,
-)
 from unilab.tasks.locomotion.common.commands import (
     Commands,
     apply_heading_yaw_feedback,
@@ -30,6 +26,10 @@ from unilab.tasks.locomotion.common.height_scan import (
     init_height_scan_sensor,
     raw_height_scan_obs,
     terrain_out_of_bounds,
+)
+from unilab.tasks.locomotion.common.terrain_spawn import (
+    TerrainCurriculumCfg,
+    TerrainSpawnManager,
 )
 from unilab.tasks.locomotion.go2w.base import NUM_GO2W_ACTIONS, NUM_LEG_ACTIONS
 from unilab.tasks.locomotion.go2w.joystick import (
